requirements.txt
fastapi
uvicorn[standard]
anthropic
httpx
pydantic
pydantic-settings
python-dotenv
duckduckgo-search
playwright
